+++ src/extension.ts
@@ -118,6 +118,10 @@ export function activate(_context: vscode.ExtensionContext) {
 	vscode.commands.registerCommand('forge.viewObjectManifest', async (object?: IObject) => {
 		await mdc.viewObjectManifest(object, context);
 	});
+	vscode.commands.registerCommand('forge.deleteObjectManifest', async (object?: IObject) => {
+		await mdc.deleteObjectManifest(object, context);
+		simpleStorageDataProvider.refresh(object);
+	});
 	vscode.commands.registerCommand('forge.viewObjectThumbnail', async (object?: IObject) => {
 		await mdc.viewObjectThumbnail(object, context);
 	});