+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added
   - Previewing and downloading of derivative thumbnails
+  - Deleting model derivatives
 
 ## [0.8.0] - 2019-09-21